Add pack support
This will add testing of the `pack` command, starting with what
was recommended on paketo website (using buildpacks.io).

Without this, one might not now where to start, where in fact,
it's bloody simple!

tox.ini had to be edited, because pack is requiring environment
variables (like HOME) to work. If hardening is requested, we can
whitelist env vars one by one. We probably should do that when
we'll have more coverage.