Add venv for commands
Without this, one needs to enter the venv to run specific command
that are installed through setuptools entrypoints.

This is mildly annoying.

This fixes it by introducing a 'venv' environment, whose purpose
is only to accept commands. Some applications need TERM or LANG
set, so we are passing _all_ the environment variables for this
venv.