Add pytest serial mark to run certain tests sequentially
We now run the tests in two stages: once all tests that are not serial in
parallel and then all the serial ones. This has the disadvantage, that for
modules where there are no serial tests, pytest will not find any tests to
execute and would fail with exit code 5 (see
https://docs.pytest.org/en/documentation-restructure/how-to/usage.html#possible-exit-codes).

To circumvent that, we use the
https://github.com/yashtodi94/pytest-custom_exit_code plugin, which allows
pytest to exit cleanly even if it did not collect any tests. However, we only
use this for the serial tests, as not collecting any tests at all could be a
bug.

This fixes https://github.com/SUSE/m8a-tests/issues/21