Restructure data
Instead of having the data inside the python data.py, we
should move to load the data from a static file, and parse
it.

The file `container.json` is an example of clean data.
Later we can complete the data, similarly to what
`container-2` currently has.

However, this means we'll have to produce another hook object.