Improve tests for format venv
Without this, the format command gets run, but it will rc 0 even
if a file needs reformatting.

This is a problem, as we might miss it.

This fixes it by skipping the formatting tox command inside the
rest of the tests, and have an extra job, for which we pass
the --check command in the CI.