Revert "Use only one instance of Docker/PodmanRuntime"
This reverts commit f3ed2a6686b022fab286b95a0204931bad9f8534.
to allow to run this tooling without podman.

Without this patch, we try to instantiate PODMAN_RUNTIME at all
times.

This is a problem, as some user might not have PODMAN, and the
program will crash with a Runtime Error.

This fixes it by only instantiating a podman runtime if it's
worth it.