Unify the tests that clone git repositories
We have a lot of tests that are essentially just `git clone $url && $do_smth`
but implemented in different places. To reduce code duplication, we add new
fixtures `container_git_clone` and `host_git_clone`, that execute the correct
git clone (supporting tags as well) inside a container or on the host and return
the test parameters.