Add performance testing
Without this, we are not comparing our BCI to UBI.
This is a problem, as we are not proving the value of our
efforts, from a technical standpoint.
We should show why our containers are probably better from
a technical perspective (or show where we can improve).

This fixes it by adding multiple builds of of the same
application, in 3 different runtimes: BCI, SLE current default
image, and UBI from RH. To make it significant, I chose
to use software which makes sense for ISVs or SUSE
Hub, so I chose Kafka.

The Kafka UBI files come from confluentinc/kafka-images/kafka.
The Dockerfile has been renamed, and follow up changes will need
to be tracked in git, so we can easily compare them.