Allow the host_git_clone fixture to work with multiple parameters
When you parametrize multiple fixtures at once, then an indirectly parametrized
one does not receive the parameter as `.param`, instead it receives a
ParameterSet object in that property which has the actual parameters as
values.
Therefore we check whether the request.param is directly what we need or whether
it's in reques.param.values[0]. And if we didn't find it there, then we error
out.