Parametrize stdout & return value of the multistage build tests
We want to actually verify that the build was successful and the resulting
container does the right thing™
To achieve that, we check that an expected string is present in the stdout of
the container. We also parametrize the return value, because maven fail when
invoked outside of a java project.