Fix build test when run on a registered SLES system
When the host system is registered, then it will automatically add the SLE
repositories to the container. This caused the check whether only one repository
is in there to fail.

Now we instead check whether we're running on SLE and whether the system is
registered. If it is, then additional repos should be added by
container-suseconnect and we check that there's more than 1. Otherwise we check
that there's just one.