Skip testing reference when TARGET is not testing released containers
We have frequently CI pipeline failures when we release a container
for the first time, as it is checking for a reference being accessible
that does not yet exist, and it would become accessible would the CI
pipeline go through. This requires manual workarounds and overriding
a release without proper testing, which isn't sensible.

As this is the only test that fails on an unreleased container, lets
skip it in that case.