Add tests for the Spack container
This is an initial implementation.

Unfortunately, a lot of the pytest_container 'goodness' was not
usable to test the Spack container:

In order to test whether sending a Spack command directly to the
container works, it needs to be launched:
 `podman run --rm <spack_container> <spack_command>`
 However, when running:
   container.connection.run_expect()
a command will be sent to a via
 `podman exec <containerID> sh -c ...`
which will not be able to run a Spack command.

Furthermore, we cannot use the MultiStageBuild.build() interface
as the simplistic substitute() replacement will choke on '$..'
shell constructs.

Co-Authored-By: Dirk Müller <dmueller@suse.com>
Signed-off-by: Egbert Eich <eich@suse.com>