Upgrade Helm to a version that doesn't have failing tests
Unfortunately this requires disabling tests on go < 1.22 because
the new version requires it. this is only temporary so I think
we can live with that.

Also limit the max processes started by go to two as otherwise
with "-n auto" parallel runs the machine can go OOM before the tests
continue as each test run is running jobs for all cores.