Disable CGO for certificate fetching test
This test compiles a go binary which performs a `http.Get`. So far, we
implicitly assumed that glibc would export the same symbols across all service
packs (the golang builder image can be based on a different service pack). This
was recently invalidated with SP6, hence we disable CGO and thereby do not rely
on glibc symbols anymore.