[appcol] Skip application collection containers on non aarch64 & x86_64
In https://github.com/SUSE/BCI-tests/pull/675 we added a new appcol python 3.11
container that uses the same test marker as the bci python 3.11 container. This
makes it impossible to filter only one of these containers which now causes
problems in QA on ppc64le and s390x because the application collection
containers are not available on those architectures. This is a quick hotfix,
that makes the QA jobs pass again, we must still fix the markers.