Wait for the stunnel container to be up before checking its logs
The `test_stunnel_logging` test was implicitly assuming that the container is
already up and running, as otherwise, there are no logs available yet. This
worked on a powerful dev machine, but fails on slow cloud worker.

Therefore, we move the healtcheck function out of the individual tests and use
it in all the stunnel test functions. This also reduces code duplication.