Skip libgcrypt tests on SLES < 15.6
On SLES < 15-SP6 libgcrypt will compute hashes for non-FIPS digests
(bsc#1229856). This will likely not be solved, because failing on those
digests is not considered a hard-requirement.

Therefore we skip the libgcrypt tests on older SLES versions, where the
behaviour of the library is to still compute hashes for those digests.